requirement.txt
appdirs==1.4.3
Django==1.9.4
django-datatable==0.2.0
django-tables2==1.5.0
mysqlclient==1.3.10
olefile==0.44
packaging==16.8
Pillow==4.1.0
pyparsing==2.2.0
six==1.10.0
uwsgi